+++ macros/src/config/for_variant.rs
@@ -1,6 +1,4 @@
 mod config;
 mod discriminant;
 
-pub(crate) use self::config::VariantConfig;
-
-use self::discriminant::DiscriminantConfig;
+pub(crate) use self::{config::*, discriminant::*};