+++ tests/derive-tests/variant_discriminant/fail/nested_on_unit_variant.rs
@@ -2,7 +2,7 @@ use enumcapsulate::VariantDiscriminant;
 
 #[derive(VariantDiscriminant)]
 pub enum Enum {
-    #[enumcapsulate(discriminant(nested = VariantADiscriminant, value = 42))]
+    #[enumcapsulate(discriminant(nested, value = 42))]
     VariantA,
 }